Move `multiplayer` package
