+++ src/com/osudroid/beatmaplisting/BeatmapDownloader.kt
@@ -6,8 +6,8 @@ import com.osudroid.resources.R.*
 import com.reco1l.framework.net.FileRequest
 import com.reco1l.framework.net.IFileRequestObserver
 import com.reco1l.osu.mainThread
-import com.reco1l.osu.multiplayer.Multiplayer
-import com.reco1l.osu.multiplayer.RoomScene
+import com.osudroid.multiplayer.Multiplayer
+import com.osudroid.multiplayer.RoomScene
 import com.osudroid.ui.v1.DownloadFragment
 import com.reco1l.toolkt.kotlin.async
 import net.lingala.zip4j.ZipFile
