+++ src/com/osudroid/ibancho/data/RoomMods.kt
@@ -1,6 +1,6 @@
 package com.osudroid.ibancho.data
 
-import com.reco1l.osu.multiplayer.Multiplayer
+import com.osudroid.multiplayer.Multiplayer
 import com.rian.osu.mods.ModCustomSpeed
 import com.rian.osu.mods.ModDifficultyAdjust
 import com.rian.osu.mods.ModDoubleTime
