+++ src/com/osudroid/multiplayer/LobbySearch.kt
@@ -1,4 +1,4 @@
-package com.reco1l.osu.multiplayer
+package com.osudroid.multiplayer
 
 import android.annotation.SuppressLint
 import android.content.Context
