+++ src/com/osudroid/multiplayer/Multiplayer.kt
@@ -1,4 +1,4 @@
-package com.reco1l.osu.multiplayer
+package com.osudroid.multiplayer
 
 import android.text.format.DateFormat
 import android.util.Log
