+++ src/com/osudroid/ui/v2/modmenu/ModMenu.kt
@@ -13,9 +13,10 @@ import com.reco1l.framework.math.*
 import com.osudroid.ibancho.RoomAPI.setPlayerMods
 import com.osudroid.ibancho.RoomAPI.setRoomMods
 import com.osudroid.ibancho.data.RoomMods
+import com.osudroid.multiplayer.Multiplayer
+import com.osudroid.multiplayer.RoomScene
 import com.osudroid.ui.OsuColors
 import com.reco1l.osu.*
-import com.reco1l.osu.multiplayer.*
 import com.osudroid.ui.v1.SettingsFragment
 import com.reco1l.toolkt.kotlin.*
 import com.reco1l.toolkt.kotlin.async
