+++ src/com/reco1l/osu/ui/entity/GameplayLeaderboard.kt
@@ -2,7 +2,7 @@ package com.reco1l.osu.ui.entity
 
 import android.opengl.GLES20
 import com.osudroid.ui.v2.hud.HUDElement
-import com.reco1l.osu.multiplayer.Multiplayer.isMultiplayer
+import com.osudroid.multiplayer.Multiplayer.isMultiplayer
 import org.anddev.andengine.entity.sprite.Sprite
 import org.anddev.andengine.entity.text.ChangeableText
 import ru.nsu.ccfit.zuev.osu.Config
