+++ src/ru/nsu/ccfit/zuev/osu/Config.java
@@ -19,7 +19,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
 
-import com.reco1l.osu.multiplayer.Multiplayer;
+import com.osudroid.multiplayer.Multiplayer;
 
 import net.margaritov.preference.colorpicker.ColorPickerPreference;
