+++ src/ru/nsu/ccfit/zuev/osu/LibraryManager.java
@@ -4,7 +4,7 @@ import static com.reco1l.osu.data.BeatmapsKt.BeatmapInfo;
 
 import android.util.Log;
 
-import com.reco1l.osu.DifficultyCalculationManager;
+import com.osudroid.DifficultyCalculationManager;
 import com.reco1l.osu.data.BeatmapSetInfo;
 import com.reco1l.osu.data.DatabaseManager;
 import com.reco1l.andengine.texture.VideoTexture;
