+++ src/ru/nsu/ccfit/zuev/osu/online/OnlineScoring.java
@@ -6,8 +6,8 @@ import android.net.Uri;
 import com.google.android.material.snackbar.Snackbar;
 import com.reco1l.osu.data.BeatmapInfo;
 import com.reco1l.osu.Execution;
-import com.reco1l.osu.multiplayer.LobbyScene;
-import com.reco1l.osu.multiplayer.RoomScene;
+import com.osudroid.multiplayer.LobbyScene;
+import com.osudroid.multiplayer.RoomScene;
 import com.rian.osu.ui.SendingPanel;
 
 import org.anddev.andengine.util.Debug;
